requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
pydantic==2.8.2
